+++ src/semaphore.h
@@ -11,23 +11,23 @@ class PseudoSemaphore
 		// fucking stupid trivial semaphore implementation, just because some moron decides semaphore is no good for high-level use!
 		// naming follows POSIX semaphore
 	public:
-		explicit PseudoSemaphore(const ssize_t value);
+		explicit PseudoSemaphore(const int value);
 
 		void waitExit();  // wait until no one is locking the semaphore
 
 		void wait();  // lock a semaphore
 		void post();   // unlock a semaphore
 
-		void setMaxValue(const ssize_t value);  // set new max value
-		ssize_t getTotalValue();  // get total value
+		void setMaxValue(const int value);  // set new max value
+		int getTotalValue();  // get total value
 
-		ssize_t getCount();  // get in-use count
-		ssize_t getValue();  // get value
+		int getCount();  // get in-use count
+		int getValue();  // get value
 
 
 	private:
-		ssize_t totalValue;
-		ssize_t nowValue;
+		int totalValue;
+		int nowValue;
 
 		std::mutex valueMtx;
 		std::condition_variable cv;
